release v0.1.3
release v0.1.3

+++ gulp/tasks/setting.js
@@ -6,40 +6,27 @@
 import path from 'path'
 import minimist from 'minimist'
 import Registry from 'undertaker-registry'
+import fs from 'fs'
 import fse from 'fs-extra'
 import request from 'request'
 import csv from 'csvtojson'
-import prettierPlugin from 'gulp-prettier-plugin'
 import config from '../config'
 
 class Setting extends Registry {
   init(gulp) {
-    const mkdirs = callback => {
-      for (const key in config.dir) {
-        fse.mkdirsSync(config.dir[key], err => {
-          if (err) {
-            console.log('errorです')
-          }
-        })
+
+    /** ファイル存在確認チェック用 **/
+    let isExist = ""
+
+    function fileExistCheck(filePath) {
+      try {
+        fs.statSync(filePath)
+        isExist = true
+      } catch (err) {
+        isExist = false
       }
 
-      for (const key in config.tools) {
-        fse.mkdirsSync(path.join(config.dir.src, config.tools[key]), err => {
-          if (err) {
-            console.log('errorです')
-          }
-        })
-      }
-
-      for (const key in config.data) {
-        fse.mkdirsSync(path.join(config.dir.src, config.tools.data, config.data[key]), err => {
-          if (err) {
-            console.log('errorです')
-          }
-        })
-      }
-
-      callback()
+      return isExist
     }
 
     const csvToJson = callback => {
@@ -48,7 +35,6 @@ class Setting extends Registry {
 
       for (const key in config.dataFile.dataname) {
         const URL = config.dataFile.dataname[key]
-        const csvFilePath = path.join(config.dataCsv, key + '.csv')
         const jsonFilePath = path.join(config.dataJson, key + '.json')
 
         csv()
@@ -63,35 +49,6 @@ class Setting extends Registry {
             count++
           }
         })
-
-        // https.get(URL, res => {
-        //   let responseString = ''
-
-        //   res.on('data', chunk => {
-        //     responseString += chunk
-        //   })
-
-        //   res.on('end', () => {
-        //     fse.outputFile(csvFilePath, responseString)
-        //     .then(() => {
-        //       csv()
-        //         .fromFile(csvFilePath)
-        //         .then((jsonObj) => {
-        //           fse.outputJson(jsonFilePath, jsonObj)
-        //         })
-        //     })
-        //     .then(() => {
-        //       if (dataLength === count) {
-        //         gulp.src(jsonFilePath)
-        //           .pipe(prettierPlugin(undefined, { filter: true }))
-        //           .pipe(gulp.dest(config.dataJson))
-        //         return callback()
-        //       } else {
-        //         count++
-        //       }
-        //     })
-        //   })
-        // })
       }
     }
 
@@ -100,11 +57,13 @@ class Setting extends Registry {
         string: 'env',
         default: {
           default: process.env.NODE_ENV || 'default/',
-          hc: process.env.NODE_ENV || 'hc/',
+          hc: process.env.NODE_ENV || 'hondacars/',
         },
       }
 
       const argv = minimist(process.argv.slice(2), args_setting)
+
+      console.log(argv)
 
       const toolsLength = Object.keys(config.tools).length - 1 //dataが不要なので−１
       let count = 1
@@ -113,41 +72,81 @@ class Setting extends Registry {
         for (let tool in config.tools) {
           if (tool === 'data') continue
 
-          fse.copy(
-            path.join(config.dir.tmp, 'default', tool),
-            path.join(config.dir.src, tool),
-            {
-              overwrite: false,
-              errorOnExist: true,
-              filter: false
-            }, () => {
-              console.log('ファイルが存在していました')
-            })
-            .then(() => {
-              console.log('template default ' + tool + ' をコピーしました')
-              if (toolsLength === count) {
-                return callback()
-              } else {
-                count++
-              }
+          fileExistCheck(path.join(config.dir.src, tool))
+          if (!isExist) {
+            fse.copy(
+                path.join(config.dir.tmp, 'default', tool),
+                path.join(config.dir.src, tool))
+              .then(() => {
+                console.log('template default ' + tool + ' をコピーしました')
+                if (toolsLength === count) {
+                  return callback()
+                } else {
+                  count++
+                }
+              })
+          } else {
+            console.log('すでにディレクトリがコピーされています')
+            if (toolsLength === count) {
+              return callback()
+            } else {
+              count++
             }
-          )
+          }
+        }
+      } else if (argv.hc === true) {
+        for (let tool in config.tools) {
+          if (tool === 'data') continue
+
+          fileExistCheck(path.join(config.dir.src, tool))
+          if (!isExist) {
+            fse.copy(
+              path.join(config.dir.tmp, argv, tool),
+              path.join(config.dir.src, tool))
+              .then(() => {
+                console.log('template hondacars' + tool + ' をコピーしました')
+                if (toolsLength === count) {
+                  return callback()
+                } else {
+                  count++
+                }
+              }
+            )
+          } else {
+            console.log('すでにディレクトリがコピーされています')
+            if (toolsLength === count) {
+              return callback()
+            } else {
+              count++
+            }
+          }
         }
       } else {
         for (let tool in config.tools) {
           if (tool === 'data') continue
-          fse.copy(
-            path.join(config.dir.tmp, argv, tool),
-            path.join(config.dir.src, tool))
-            .then(() => {
-              console.log('template default ' + tool + ' をコピーしました')
-              if (toolsLength === count) {
-                return callback()
-              } else {
-                count++
+
+          fileExistCheck(path.join(config.dir.src, tool))
+          if (!isExist) {
+            fse.copy(
+              path.join(config.dir.tmp, 'default', tool),
+              path.join(config.dir.src, tool))
+              .then(() => {
+                console.log('template default ' + tool + ' をコピーしました')
+                if (toolsLength === count) {
+                  return callback()
+                } else {
+                  count++
+                }
               }
+            )
+          } else {
+            console.log('すでにディレクトリがコピーされています')
+            if (toolsLength === count) {
+              return callback()
+            } else {
+              count++
             }
-          )
+          }
         }
       }
     }
@@ -155,30 +154,53 @@ class Setting extends Registry {
     const fileSet = callback => {
       const sitemapDataFile = path.join(config.dataJson, 'sitemap.json')
       const sitemapData = JSON.parse(fse.readFileSync(sitemapDataFile, 'utf8'))
+      const ejsData = fs.readFileSync(path.join(config.dir.tmp, 'template.ejs'), 'utf-8')
+      const ejsScss = fs.readFileSync(path.join(config.dir.tmp, 'template.scss'), 'utf-8')
 
       for (const pages in sitemapData) {
-        fse.outputFileSync(
-          path.join(config.dir.src, config.tools.ejs, sitemapData[pages].page + '.ejs'),
-          ''
-        )
+        if ( sitemapData[pages].page === 'index' ){
+          fileExistCheck(path.join(config.dir.src, config.tools.ejs, sitemapData[pages].page + '.ejs'))
 
-        fse.outputFileSync(
-          path.join(config.dir.src, config.tools.sass, sitemapData[pages].page + '.scss'),
-          ''
-        )
+          if (!isExist){
+            fse.outputFileSync(
+              path.join(config.dir.src, config.tools.ejs, sitemapData[pages].page + '.ejs'),
+              ejsData
+            )
+          } else {
+            console.log('すでにファイルが存在していました')
+          }
+        } else {
+          fileExistCheck(path.join(config.dir.src, config.tools.ejs, sitemapData[pages].category, sitemapData[pages].page + '.ejs'))
+
+          if (!isExist) {
+            fse.outputFileSync(
+              path.join(config.dir.src, config.tools.ejs, sitemapData[pages].category, sitemapData[pages].page + '.ejs'),
+              ejsData
+            )
+          } else {
+            console.log('すでにファイルが存在していました')
+          }
+        }
+
+        fileExistCheck(path.join(config.dir.src, config.tools.sass, sitemapData[pages].page + '.scss'))
+        if (!isExist) {
+          fse.outputFileSync(
+            path.join(config.dir.src, config.tools.sass, 'project', sitemapData[pages].page + '.scss'), ejsScss
+          )
+        } else {
+          console.log('すでにファイルが存在していました')
+        }
       }
 
       callback()
     }
 
-    gulp.task('mkdirs', mkdirs)
     gulp.task('csvToJson', csvToJson)
     gulp.task('fileCopy', fileCopy)
     gulp.task('fileSet', fileSet)
 
     const fileCreate = callback => {
       gulp.series(
-        `mkdirs`,
         `csvToJson`,
         `fileCopy`,
         'fileSet',
@@ -189,7 +211,6 @@ class Setting extends Registry {
       )()
 
       // TODO json整形タスクも追加する
-
       callback()
     }
